+++ azimuth/modules/model_performance/confusion_matrix.py
@@ -6,6 +6,8 @@ from typing import List
 
 import numpy as np
 from datasets import Dataset
+from scipy.sparse import csr_matrix
+from scipy.sparse.csgraph import reverse_cuthill_mckee
 from sklearn.metrics import confusion_matrix
 
 from azimuth.config import ModelContractConfig
@@ -13,11 +15,16 @@ from azimuth.modules.base_classes import FilterableModule
 from azimuth.types.model_performance import ConfusionMatrixResponse
 from azimuth.utils.validation import assert_not_none
 
+MIN_CONFUSION_CUTHILL_MCKEE = 0.1
+
 
 class ConfusionMatrixModule(FilterableModule[ModelContractConfig]):
     """Computes the confusion matrix on the specified dataset split."""
 
-    allowed_mod_options = FilterableModule.allowed_mod_options | {"cf_normalized"}
+    allowed_mod_options = FilterableModule.allowed_mod_options | {
+        "cf_normalized",
+        "cf_preserved_class_order",
+    }
 
     def compute_on_dataset_split(self) -> List[ConfusionMatrixResponse]:  # type: ignore
         """Computes confusion matrix from sklearn.
@@ -34,6 +41,9 @@ class ConfusionMatrixModule(FilterableModule[ModelContractConfig]):
         ds_mng = self.get_dataset_split_manager()
         num_classes = ds_mng.get_num_classes()
         class_ids = list(range(num_classes))
+        class_names = ds_mng.get_class_names()
+        rejection_idx = ds_mng.rejection_class_idx
+
         cf = confusion_matrix(
             y_true=labels,
             y_pred=predictions,
@@ -41,19 +51,43 @@ class ConfusionMatrixModule(FilterableModule[ModelContractConfig]):
             normalize="true" if self.mod_options.cf_normalized else None,
         )
 
-        class_names = ds_mng.get_class_names()
+        # Reorder rows and columns so the bandwidth of the matrix is smaller
+        if not self.mod_options.cf_preserved_class_order:
+            # Get a normalized confusion matrix if not already computed
+            if not self.mod_options.cf_normalized:
+                cf_normalized = confusion_matrix(
+                    y_true=labels, y_pred=predictions, labels=class_ids, normalize="true"
+                )
+            else:
+                cf_normalized = cf
 
-        # Put the rejection class last for the confusion matrix
-        rejection_idx = ds_mng.rejection_class_idx
-        if rejection_idx != num_classes - 1:
-            new_order = class_ids[:rejection_idx] + class_ids[rejection_idx + 1 :] + [rejection_idx]
-            cf = cf[np.ix_(new_order, new_order)]
-            class_names = [class_names[i] for i in new_order]
+            # Replace values of the rejection class by 0 so they don't influence the algorithm
+            cf_rej_class_0 = cf_normalized.copy()
+            cf_rej_class_0[rejection_idx] = 0
+            cf_rej_class_0[:, rejection_idx] = 0
+
+            # Get order based on reverse_cuthill_mckee algorithm
+            graph = csr_matrix(cf_rej_class_0 >= MIN_CONFUSION_CUTHILL_MCKEE)
+            order = reverse_cuthill_mckee(graph)
+
+            # Put the rejection class last for the confusion matrix
+            if rejection_idx != order[-1]:
+                rejection_position = np.where(order == rejection_idx)[0][0]
+                # From the rejection position, roll backwards all values
+                order[rejection_position:] = np.roll(order[rejection_position:], -1)
+
+            # Re-order matrix
+            cf = cf[np.ix_(order, order)]
+            class_names = [ds_mng.get_class_names()[i] for i in order]
 
         return [
             ConfusionMatrixResponse(
                 confusion_matrix=cf,
                 class_names=class_names,
                 normalized=self.mod_options.cf_normalized,
+                preserved_class_order=self.mod_options.cf_preserved_class_order,
+                rejection_class_position=rejection_idx
+                if self.mod_options.cf_preserved_class_order
+                else num_classes - 1,
             )
         ]