+++ azimuth/routers/v1/model_performance/confusion_matrix.py
@@ -40,12 +40,14 @@ def get_confusion_matrix(
     pipeline_index: int = Depends(require_pipeline_index),
     without_postprocessing: bool = Query(False, title="Without Postprocessing"),
     normalized: bool = Query(True, title="Normalized"),
+    preserved_class_order: bool = Query(False, title="Preserved Class Order"),
 ) -> ConfusionMatrixResponse:
     mod_options = ModuleOptions(
         filters=named_filters.to_dataset_filters(dataset_split_manager.get_class_names()),
         pipeline_index=pipeline_index,
         without_postprocessing=without_postprocessing,
         cf_normalized=normalized,
+        cf_preserved_class_order=preserved_class_order,
     )
 
     task_result: ConfusionMatrixResponse = get_standard_task_result(